requirements.txt
# Core Machine Learning Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Computer Vision Libraries
opencv-python>=4.8.0
Pillow>=10.0.0
mtcnn>=0.1.1

# Deep Learning and AI
transformers>=4.31.0
efficientnet-pytorch>=0.7.1
albumentations>=1.3.1
timm>=0.9.2

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0

# Web Application
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Utilities
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0.0

# Logging
loguru>=0.7.0

# Progress Bars
rich>=13.4.0

# Memory Management
psutil>=5.9.0

# Model Serialization
joblib>=1.3.0

# Image Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Statistical Analysis
scipy>=1.11.0

# Natural Language Processing (for sentiment analysis)
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0
